Checkpoint: Full functional web app with 3D Orbiting cards, floating glass navbar, API wiring, and responsive layout

+++ Backend/api/controllers/user.controller.js
@@ -2,16 +2,27 @@ import User from "../models/user.model.js";
 import createError from "../utils/createError.js";
 
 export const deleteUser = async (req, res, next) => {
-  const user = await User.findById(req.params.id);
+  try {
+    const user = await User.findById(req.params.id);
+    if (!user) return next(createError(404, "User not found!"));
 
-  if (req.userId !== user._id.toString()) {
-    return next(createError(403, "You can delete only your account!"));
+    if (req.userId !== user._id.toString()) {
+      return next(createError(403, "You can delete only your account!"));
+    }
+    await User.findByIdAndDelete(req.params.id);
+    res.status(200).send("deleted.");
+  } catch (err) {
+    next(err);
   }
-  await User.findByIdAndDelete(req.params.id);
-  res.status(200).send("deleted.");
 };
-export const getUser = async (req, res, next) => {
-  const user = await User.findById(req.params.id);
 
-  res.status(200).send(user);
+export const getUser = async (req, res, next) => {
+  try {
+    const user = await User.findById(req.params.id);
+    if (!user) return next(createError(404, "User not found!"));
+    const { password, ...rest } = user._doc;
+    res.status(200).send(rest);
+  } catch (err) {
+    next(err);
+  }
 };
